refactor(doctor): extract engine.rs, dispatch fix via Check::fix trait method
